Extended the Interface
Added a Status flag in the resolver so clients can know if a scan is in
progress, and throw an exception if a Scan is initiated while a scan is
in progress. Added an indication of when scan is completed via the
callback. Test and handle cancelling scan more gracefully.